requirements.txt
jira>=3.10.5
python-dotenv>=1.2.2
